analysis/requirements.txt
pymongo
polars
python-dotenv

